+++ src/config/settings/settings.mjs
@@ -1121,6 +1121,10 @@ const SETTINGS = {
             type: Boolean,
             default: false,
           },
+          "munching-policy-remove-weapon-mastery-description": {
+            type: Boolean,
+            default: false,
+          },
         },
         ENCOUNTER: {
           "encounter-import-policy-create-scene": {